# hyperf/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "hyperf swagger api desc",
        "version": "1.0.0",
        "title": "HYPERF API DOC"
    },
    "host": "39.105.46.139:9503",
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/project/info": {
            "post": {
                "tags": [
                    "项目管理"
                ],
                "summary": "",
                "description": "项目详情",
                "operationId": "ProjectInfoPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "project_id",
                        "description": "项目id",
                        "required": true,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
...
        "/user/project": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "",
                "description": "我的项目列表",
                "operationId": "UserProjectPOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "status",
                        "description": "状态",
                        "required": true,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/user/update": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "",
                "description": "我的项目列表",
                "operationId": "UserUpdatePOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "密码",
                        "required": false,
                        "type": "string",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "address",
                        "description": "常驻地址",
                        "required": false,
                        "type": "string",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "real_name",
                        "description": "真实姓名",
                        "required": false,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/user/change_new_phone": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "",
                "description": "修改手机号码",
                "operationId": "UserChange_new_phonePOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "new_phone",
                        "description": "新手机号码",
                        "required": true,
                        "type": "string",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "new_phone_code",
                        "description": "新手机验证码",
                        "required": true,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/user/check_old_phone": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "",
                "description": "检测老手机号码",
                "operationId": "UserCheck_old_phonePOST",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "old_phone",
                        "description": "原手机号码",
                        "required": true,
                        "type": "string",
                        "default": null
                    },
                    {
                        "in": "formData",
                        "name": "old_phone_code",
                        "description": "原手机验证码",
                        "required": true,
                        "type": "string",
                        "default": null
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": [],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
...
    }
}